Size distribution analysis addition started Fixes #377
